BackEnd/requirements.txt
asgiref==3.7.2
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
cryptography==42.0.3
defusedxml==0.7.1
Django==4.2.10
django-allauth==0.55.0
django-cors-headers==4.3.1
djangorestframework==3.14.0
drf-yasg==1.21.7
idna==3.6
inflection==0.5.1
oauthlib==3.2.2
packaging==23.2
pycparser==2.21
PyJWT==2.8.0
python3-openid==3.2.0
pytz==2024.1
PyYAML==6.0.1
requests==2.31.0
requests-oauthlib==1.3.1
setuptools==65.5.1
sqlparse==0.4.4
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.0
wheel==0.38.4